rename to intersect
