+++ core/amber/src/test/scala/edu/uci/ics/texera/unittest/workflow/operators/intersection/IntersectOpExecSpec.scala
@@ -4,7 +4,7 @@ import edu.uci.ics.amber.engine.common.InputExhausted
 import edu.uci.ics.amber.engine.common.virtualidentity.{LayerIdentity, LinkIdentity}
 import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType}
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
-import edu.uci.ics.texera.workflow.operators.intersection.IntersectOpExec
+import edu.uci.ics.texera.workflow.operators.intersect.IntersectOpExec
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.BeforeAndAfter
